Add optionContent spec coverage
